requirements.txt
python-telegram-bot==20.7
ffmpeg-python==0.2.0